Move ForcedWithdrawRepository
