+++ packages/backend/src/peripherals/database/transactions/ForcedWithdrawRepository.test.ts
@@ -1,8 +1,8 @@
 import { Hash256, StarkKey, Timestamp } from '@explorer/types'
 import { expect } from 'earljs'
 
-import { setupDatabaseTestSuite } from '../../test/database'
-import { Logger } from '../../tools/Logger'
+import { setupDatabaseTestSuite } from '../../../test/database'
+import { Logger } from '../../../tools/Logger'
 import {
   ForcedWithdrawRepository,
   ForcedWithdrawTransactionRecord,
